- Columns UpdatedAtTrait.php and CreatedAtTrait.php seperated into separated traits as it should have been.
